Fixed a minor bug in vaccine_compliance. Fixed the VaccineTrace file to print the right format Updated the vaccine regression test.
git-svn-id: https://redmine.hindsight.psc.edu/svn/fred/trunk/FRED@205 b93b844c-2574-459d-9dfb-c4b8cb2af1a1

+++ Vaccine_Manager.h
@@ -49,7 +49,7 @@ public:
   void vaccinate(int day);
   
   //Paramters Access Members
-  int get_vaccine_compliance()        const {return vaccine_compliance;}
+  double get_vaccine_compliance()        const {return vaccine_compliance;}
   int get_vaccine_priority_age_low()  const {return vaccine_priority_age_low;}
   int get_vaccine_priority_age_high() const {return vaccine_priority_age_high;}
